Upgrade to AndroidStudio 4.0 Canary 1 and AGP 4.0.0-alpha01
Bug: b/142859295
Fixes: b/143452527
Test: ./studiow, ./gradlew boS

+++ ui/ui-android-text/src/main/java/androidx/text/style/BaselineShiftSpan.java
@@ -42,12 +42,12 @@ public class BaselineShiftSpan extends MetricAffectingSpan {
 
     @Override
     public void updateMeasureState(@NonNull TextPaint textPaint) {
-        textPaint.baselineShift += textPaint.ascent() * mMultiplier;
+        textPaint.baselineShift += (int)(textPaint.ascent() * mMultiplier);
     }
 
     @Override
     public void updateDrawState(@NonNull TextPaint textPaint) {
-        textPaint.baselineShift += textPaint.ascent() * mMultiplier;
+        textPaint.baselineShift += (int)(textPaint.ascent() * mMultiplier);
     }
 
     public float getMultiplier() {
